+++ src/pages/dashboard/TeamHelp/TeamHelp.scss
@@ -1,24 +1,24 @@
 @import '../../../css/variables';
 
-.mainWorkflowCard {
+.mainTeamCard {
 	border-top: 1px solid #e7e9ed !important;
 }
 
-.workflowSlide {
+.teamSlide {
 	margin-bottom: 16px;
 }
 
-.workflowAnswer {
+.teamAnswer {
     margin-left: 29px;
 }
 
-.workflowQuestion {
+.teamQuestion {
     margin-left: 0px;
     margin-top: 12px;
     margin-bottom: 12px;
 }
 
-.workflowRectangle {
+.teamRectangle {
     width: 100%;
     box-shadow: 1px 1px 3px 0 #00000017;
     background-color: $white;
@@ -26,6 +26,6 @@
     padding-bottom: 8px;
   }
 
-.workflowSubtitle {
+.teamSubtitle {
     padding-bottom: 8px;
 }
